backend/requirements.txt
Django==5.0.2
djangorestframework
django-storages
Pillow
django-cors-headers==4.3.1
python-dotenv
djangorestframework-simplejwt
psycopg2-binary
django-filter==23.3
sqlparse==0.4.4
six
django-taggit
moviepy